Updates the method for creating the popup window (now uses chrome.windows.create) Using new method to close the popup window since the delicious.com redesign.
We now send the ID of the popup window the extension created, to the
onUpdated listener. We then check for that ID via the onUpdated
listener, and if it's URL does not 'start with' delcious.com/save then
it's safe to close our window. A bit janky, I assume, but it's doing
the trick. Also, commented out the options that were previously
available when you clicked the extensions icon. Instead, it just
initiates the 'save' popup.