No longer closes popup if you're not logged in. Push to login page. Adjusts layout of login page so it fits in our popup.
No longer closing our popup if you're not logged in when you initiate
the extensions popup. Push to the login page instead of the default,
delicious.com/register page (chances are, if you have this extension,
you are already registered). Adjusts layout of login page so it fits in
our popup (style.css).